Revert "A new hope summer edition"
