+++ app/src/main/java/br/edu/insper/al/grupo_8/projeto_2/MethodPacienteActivity.java
@@ -38,8 +38,8 @@ public class MethodPacienteActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_method_paciente);
 
-        Button buttonTests = findViewById(R.id.button_goMenu);
-        Button buttonGoHome = findViewById(R.id.button_goBack);
+        Button buttonTests = findViewById(R.id.button_goTests);
+        Button buttonGoHome = findViewById(R.id.button_goLog);
 
         TextView dataIn = findViewById(R.id.data_txt);
         EditText nomeIn = findViewById(R.id.nome_txt);
