+++ app/src/main/java/br/edu/insper/al/grupo_8/projeto_2/methods/MethodVIActivity.java
@@ -19,7 +19,6 @@ import br.edu.insper.al.grupo_8.projeto_2.R;
 import br.edu.insper.al.grupo_8.projeto_2.TestsActivity;
 
 public class MethodVIActivity extends AppCompatActivity {
-    private String rh;
     private HashMap<String, String> testeF = new HashMap<String, String>();
 
     //METODO 1
@@ -143,9 +142,17 @@ public class MethodVIActivity extends AppCompatActivity {
     private LinkedHashMap<Integer, HashMap> TESTE3 = new LinkedHashMap<>();
     private LinkedHashMap<Integer, HashMap> TESTEF = new LinkedHashMap<>();
 
+
+
+
+
+
+
+
+
+
     private void startMethodActivity(Class classe) {
         Intent intent = new Intent(this, classe);
-        intent.putExtra("rh", rh);
         startActivity(intent);
     }
 
@@ -154,14 +161,13 @@ public class MethodVIActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_method_vi);
 
-        Bundle extras = getIntent().getExtras();
-        assert extras != null;
-        rh = extras.getString("rh");
+        Button buttonGoBack = findViewById(R.id.button_goTests);
+        buttonGoBack.setOnClickListener((view) -> startMethodActivity(MethodVActivity.class));
 
         Button buttonGoMenu = findViewById(R.id.button_goMenu);
         buttonGoMenu.setOnClickListener((view) -> startMethodActivity(TestsActivity.class));
 
-        Button buttonNext = findViewById(R.id.button_goMethodVII);
+        Button buttonNext = findViewById(R.id.button_goMethodIII);
         buttonNext.setOnClickListener((view) -> startMethodActivity(MethodVIIActivity.class));
 
         quanto_essa_medicacao_a1 = (EditText) findViewById(R.id.quanto_essa_medicacao_a1);
