--- app/src/main/res/layout-sw600dp/activity_home.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".HomeActivity">
-    
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/green_unifesp"
-        android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <Button
-                android:id="@+id/button_goOut"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:background="@drawable/custom_buttom"
-                android:text="Logoff"
-                android:layout_alignParentStart="true"
-                android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
-                android:src="@drawable/unifesp_escrito" />
-        </RelativeLayout>
-    </android.support.v7.widget.Toolbar>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_centerInParent="true"
-        android:layout_marginStart="250dp"
-        android:gravity="center">
-
-        <Button
-            android:id="@+id/add_pacient"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:paddingBottom="60dp"
-            android:paddingTop="60dp"
-            android:paddingLeft="20dp"
-            android:paddingRight="20dp"
-            android:layout_marginEnd="30dp"
-            android:background="@drawable/custom_buttom"
-            android:textColor="@color/green_unifesp"
-            android:textSize="18sp"
-            android:textAllCaps="true"
-            android:textStyle="bold"
-            android:text="Adicionar Paciente" />
-
-        <ListView
-            android:id="@+id/pacients"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:choiceMode="singleChoice" />
-    </LinearLayout>
-</RelativeLayout>
-
