--- app/src/main/res/layout-sw600dp/activity_log_in.xml
@@ -1,114 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/green_unifesp"
-    android:orientation="vertical"
-    tools:context=".LogInActivity">
-
-    <LinearLayout android:id="@+id/content"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <ImageView
-            android:id="@+id/unifesp_logo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:contentDescription="unifesp_logo"
-            android:src="@drawable/unifesp_marca" />
-
-        <LinearLayout
-            android:id="@+id/login"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Login: "
-                android:textColor="@color/white"
-                android:textSize="33sp" />
-
-            <android.support.v7.widget.AppCompatEditText
-                android:id="@+id/text_login"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:ems="13"
-                android:inputType="textEmailAddress"
-                android:textColor="@color/white"
-                android:textColorHint="@color/white"
-                android:textCursorDrawable="@null"
-                android:textSize="25sp"
-                app:backgroundTint="@color/white" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/password"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Senha: "
-                android:textColor="@color/white"
-                android:textSize="33sp" />
-
-            <android.support.v7.widget.AppCompatEditText
-                android:id="@+id/text_password"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:ems="13"
-                android:inputType="textPassword"
-                android:textColor="@color/white"
-                android:textCursorDrawable="@null"
-                android:textSize="25sp"
-                app:backgroundTint="@color/white" />
-        </LinearLayout>
-
-        <Button
-            android:id="@+id/button_login"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="50sp"
-            android:background="@drawable/custom_buttom"
-            android:text="Entrar"
-            android:textColor="@color/green_unifesp"
-            android:textSize="20sp"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/button_add"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15sp"
-            android:background="@drawable/custom_buttom"
-            android:text="Adicionar novo Médico"
-            android:textColor="@color/green_unifesp"
-            android:textSize="20sp"
-            android:textStyle="bold" />
-    </LinearLayout>
-
-    <ProgressBar
-        android:id="@+id/progressBar"
-        style="?android:attr/progressBarStyle"
-        android:layout_marginTop="20sp"
-        android:layout_below="@+id/content"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:visibility="gone"/>
-</RelativeLayout>
