--- app/src/main/res/layout-sw600dp/activity_pacient_log.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/green_unifesp"
-        android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <Button
-                android:id="@+id/button_goBack"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:background="@drawable/custom_buttom"
-                android:text="Voltar"
-                android:layout_alignParentStart="true"
-                android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
-                android:src="@drawable/unifesp_escrito" />
-        </RelativeLayout>
-    </android.support.v7.widget.Toolbar>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:id="@+id/nome"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/black"
-            android:layout_centerHorizontal="true"
-            android:textSize="25dp"
-            android:textStyle="bold"/>
-
-        <Button
-            android:id="@+id/button_results"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@drawable/custom_buttom"
-            android:layout_above="@+id/button_tests"
-            android:layout_centerHorizontal="true"
-            android:paddingTop="40dp"
-            android:paddingBottom="40dp"
-            android:layout_marginStart="500dp"
-            android:layout_marginBottom="20dp"
-            android:textColor="@color/green_unifesp"
-            android:textStyle="bold"
-            android:textSize="20sp"
-            android:text="RESULTADOS" />
-
-        <Button
-            android:id="@+id/button_tests"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:layout_marginStart="500dp"
-            android:layout_marginBottom="20dp"
-            android:background="@drawable/custom_buttom"
-            android:paddingTop="40dp"
-            android:paddingBottom="40dp"
-            android:text="FAZER TESTES"
-            android:textColor="@color/green_unifesp"
-            android:textSize="20sp"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/button_delete"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/button_tests"
-            android:background="@drawable/custom_buttom"
-            android:layout_centerHorizontal="true"
-            android:paddingTop="40dp"
-            android:paddingBottom="40dp"
-            android:layout_marginStart="500dp"
-            android:textColor="@color/green_unifesp"
-            android:textStyle="bold"
-            android:textSize="20sp"
-            android:text="EXCLUIR PACIENTE" />
-    </RelativeLayout>
-</LinearLayout>
