--- app/src/main/res/layout-sw600dp/activity_profile.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ProfileActivity"
-    android:background="#27693F">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/green_unifesp"
-        android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
-
-        <ImageView
-            android:layout_width="146dp"
-            android:layout_height="45dp"
-            android:layout_gravity="center"
-            android:contentDescription="Unifesp_Escrito"
-            android:src="@drawable/unifesp_escrito" />
-    </android.support.v7.widget.Toolbar>
-
-    <LinearLayout android:id="@+id/content"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/toolbar"
-        android:orientation="vertical"
-        android:layout_centerHorizontal="true">
-
-        <LinearLayout
-            android:id="@+id/login"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Login: "
-                android:textColor="@color/white"
-                android:textSize="33sp" />
-
-            <android.support.v7.widget.AppCompatEditText
-                android:id="@+id/text_login"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:ems="13"
-                android:inputType="textEmailAddress"
-                android:textColor="@color/white"
-                android:textColorHint="@color/white"
-                android:textCursorDrawable="@null"
-                android:textSize="25sp"
-                app:backgroundTint="@color/white" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/password"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:orientation="horizontal">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Senha: "
-                android:textColor="@color/white"
-                android:textSize="33sp" />
-
-            <android.support.v7.widget.AppCompatEditText
-                android:id="@+id/text_password"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:ems="13"
-                android:inputType="textPassword"
-                android:textColor="@color/white"
-                android:textCursorDrawable="@null"
-                android:textSize="25sp"
-                app:backgroundTint="@color/white" />
-        </LinearLayout>
-
-        <Button
-            android:id="@+id/button_add"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="50sp"
-            android:background="@drawable/custom_buttom"
-            android:text="Adicionar novo médico"
-            android:textColor="@color/green_unifesp"
-            android:textSize="20sp"
-            android:textStyle="bold" />
-    </LinearLayout>
-
-    <ProgressBar
-        android:id="@+id/progressBar"
-        android:layout_below="@+id/content"
-        style="?android:attr/progressBarStyle"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="20dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:visibility="gone"/>
-</RelativeLayout>
