+++ app/src/main/res/layout/activity_home.xml
@@ -12,28 +12,29 @@
         android:layout_height="wrap_content"
         android:background="@color/green_unifesp"
         android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
+        android:theme="?attr/actionBarTheme">
 
         <RelativeLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
 
             <Button
                 android:id="@+id/button_goOut"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:background="@drawable/custom_buttom"
                 android:text="Logoff"
-                android:layout_alignParentEnd="true"
-                android:background="@color/green_unifesp"
-                android:textColor="#fff"
                 android:textAllCaps="true"
-                android:textSize="14sp"
+                android:textSize="18sp"
                 android:textStyle="bold" />
 
             <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:contentDescription="Unifesp_Escrito"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:contentDescription="unifesp_escrito"
+                android:layout_centerInParent="true"
                 android:src="@drawable/unifesp_escrito" />
         </RelativeLayout>
     </android.support.v7.widget.Toolbar>
@@ -41,22 +42,21 @@
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_below="@+id/toolbar"
-        android:layout_marginTop="15dp"
-        android:layout_centerInParent="true">
+        android:orientation="horizontal"
+        android:layout_centerInParent="true"
+        android:layout_marginStart="250dp"
+        android:gravity="center"
+        tools:ignore="UselessParent">
 
         <Button
             android:id="@+id/add_pacient"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:paddingBottom="30dp"
-            android:paddingTop="30dp"
-            android:paddingLeft="10dp"
-            android:paddingRight="10dp"
-            android:layout_marginEnd="15dp"
-            android:layout_marginStart="15dp"
-            android:layout_marginBottom="15dp"
+            android:paddingBottom="60dp"
+            android:paddingTop="60dp"
+            android:paddingLeft="20dp"
+            android:paddingRight="20dp"
+            android:layout_marginEnd="30dp"
             android:background="@drawable/custom_buttom"
             android:textColor="@color/green_unifesp"
             android:textSize="18sp"
