+++ app/src/main/res/layout/activity_log_in.xml
@@ -18,8 +18,8 @@
 
         <ImageView
             android:id="@+id/unifesp_logo"
-            android:layout_width="300dp"
-            android:layout_height="180dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
             android:layout_gravity="center"
             android:contentDescription="unifesp_logo"
             android:src="@drawable/unifesp_marca" />
@@ -29,7 +29,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
-            android:layout_marginTop="3dp"
+            android:layout_marginTop="30dp"
             android:orientation="horizontal">
 
             <TextView
@@ -37,18 +37,18 @@
                 android:layout_height="wrap_content"
                 android:text="Login: "
                 android:textColor="@color/white"
-                android:textSize="25sp" />
+                android:textSize="33sp" />
 
             <android.support.v7.widget.AppCompatEditText
                 android:id="@+id/text_login"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:ems="8"
+                android:ems="13"
                 android:inputType="textEmailAddress"
                 android:textColor="@color/white"
                 android:textColorHint="@color/white"
                 android:textCursorDrawable="@null"
-                android:textSize="22sp"
+                android:textSize="25sp"
                 app:backgroundTint="@color/white" />
         </LinearLayout>
 
@@ -57,7 +57,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
-            android:layout_marginTop="3dp"
+            android:layout_marginTop="30dp"
             android:orientation="horizontal">
 
             <TextView
@@ -65,47 +65,47 @@
                 android:layout_height="wrap_content"
                 android:text="Senha: "
                 android:textColor="@color/white"
-                android:textSize="25sp" />
+                android:textSize="33sp" />
 
             <android.support.v7.widget.AppCompatEditText
                 android:id="@+id/text_password"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:ems="8"
+                android:ems="13"
                 android:inputType="textPassword"
                 android:textColor="@color/white"
                 android:textCursorDrawable="@null"
-                android:textSize="22sp"
+                android:textSize="25sp"
                 app:backgroundTint="@color/white" />
         </LinearLayout>
 
         <Button
             android:id="@+id/button_login"
-            android:layout_width="250dp"
+            android:layout_width="300dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="30sp"
+            android:layout_marginTop="50sp"
             android:background="@drawable/custom_buttom"
             android:text="Entrar"
             android:textColor="@color/green_unifesp"
-            android:textSize="18sp"
+            android:textSize="20sp"
             android:textStyle="bold" />
 
         <Button
             android:id="@+id/button_add"
-            android:layout_width="250dp"
+            android:layout_width="300dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="15sp"
             android:background="@drawable/custom_buttom"
             android:text="Adicionar novo Médico"
             android:textColor="@color/green_unifesp"
-            android:textSize="18sp"
+            android:textSize="20sp"
             android:textStyle="bold" />
     </LinearLayout>
 
     <ProgressBar
         android:id="@+id/progressBar"
-        style="@android:style/Widget.ProgressBar"
-        android:layout_marginTop="10sp"
+        style="?android:attr/progressBarStyle"
+        android:layout_marginTop="20sp"
         android:layout_below="@+id/content"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
