+++ app/src/main/res/layout/activity_method_ii.xml
@@ -8,46 +8,94 @@
     android:orientation="vertical">
 
     <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
+        android:id="@+id/toolbar2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/green_unifesp"
         android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
+        android:theme="?attr/actionBarTheme">
 
-        <RelativeLayout
+        <GridLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            android:columnCount="50"
+            android:orientation="vertical"
+            android:rowCount="1">
+
+            <Button
+                android:id="@+id/button_goTests"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+
+                android:background="@drawable/custom_buttom"
+
+                android:text="Voltar"
+                android:textAllCaps="true"
+                android:textSize="18dp"
+                android:textStyle="bold"
+                tools:layout_conversion_absoluteHeight="0dp"
+                tools:layout_editor_absoluteY="88dp"
+                tools:layout_editor_absoluteX="265dp"
+                tools:layout_conversion_absoluteWidth="0dp"
+                android:layout_column="0"/>
+
+            <Space
+                android:id="@+id/Space"
+                android:layout_width="202dp"
+                android:layout_height="match_parent"
+                android:layout_column="1" />
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="60dp"
+                android:layout_row="0"
+                android:layout_column="46"
+                android:contentDescription="unifesp_escrito"
+                android:src="@drawable/unifesp_escrito"
+                tools:layout_conversion_absoluteHeight="0dp"
+                tools:layout_conversion_absoluteWidth="0dp"
+                tools:layout_editor_absoluteX="265dp"
+                tools:layout_editor_absoluteY="88dp" />
 
             <Button
                 android:id="@+id/button_goMenu"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+
                 android:background="@drawable/custom_buttom"
                 android:text="Menu"
-                android:layout_alignParentStart="true"
                 android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
-
-            <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
-                android:src="@drawable/unifesp_escrito" />
+                android:textSize="18dp"
+                android:textStyle="bold"
+                tools:layout_conversion_absoluteHeight="0dp"
+                tools:layout_editor_absoluteY="88dp"
+                tools:layout_editor_absoluteX="265dp"
+                tools:layout_conversion_absoluteWidth="0dp"
+                android:layout_row="0"
+                android:layout_column="48" />
 
             <Button
                 android:id="@+id/button_goMethodIII"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
+
                 android:background="@drawable/custom_buttom"
-                android:text="Salvar e Continuar"
+
+                android:text="Próximo"
                 android:textAllCaps="true"
                 android:textSize="18dp"
-                android:textStyle="bold" />
-        </RelativeLayout>
+                android:textStyle="bold"
+                tools:layout_conversion_absoluteHeight="0dp"
+                tools:layout_editor_absoluteY="88dp"
+                tools:layout_editor_absoluteX="265dp"
+                tools:layout_conversion_absoluteWidth="0dp"
+                android:layout_row="0"
+                android:layout_column="49" />
+
+
+
+        </GridLayout>
+
     </android.support.v7.widget.Toolbar>
 
     <Space
