+++ app/src/main/res/layout/activity_method_iii.xml
@@ -8,46 +8,76 @@
     tools:context=".methods.MethodIIIActivity">
 
     <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
+        android:id="@+id/toolbar2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/green_unifesp"
         android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
+        android:theme="?attr/actionBarTheme">
 
-        <RelativeLayout
+        <GridLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            android:columnCount="50"
+            android:rowCount="1">
 
             <Button
-                android:id="@+id/button_goMenu"
+                android:id="@+id/button_goTests"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:layout_gravity="end"
+
                 android:background="@drawable/custom_buttom"
-                android:text="Menu"
-                android:layout_alignParentStart="true"
+
                 android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
+                android:textSize="18dp"
+                android:textStyle="bold"
+                android:text="Voltar"/>
 
             <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_gravity="center"
+                android:layout_marginLeft="210dp"
+                android:contentDescription="unifesp_escrito"
                 android:src="@drawable/unifesp_escrito" />
 
             <Button
-                android:id="@+id/button_goMethodIV"
-                android:layout_width="wrap_content"
+                android:id="@+id/button_goMenu"
+                android:layout_width="110dp"
                 android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
+                android:layout_row="0"
+
+                android:layout_column="48"
+
+                android:layout_gravity="end"
                 android:background="@drawable/custom_buttom"
-                android:text="Salvar e Continuar"
+                android:text="Menu"
                 android:textAllCaps="true"
                 android:textSize="18dp"
                 android:textStyle="bold" />
-        </RelativeLayout>
+
+            <Button
+                android:id="@+id/button_goMethodIII"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+
+                android:background="@drawable/custom_buttom"
+
+                android:text="Próximo"
+                android:textAllCaps="true"
+                android:textSize="18dp"
+                android:textStyle="bold"
+                tools:layout_conversion_absoluteHeight="0dp"
+                tools:layout_editor_absoluteY="88dp"
+                tools:layout_editor_absoluteX="265dp"
+                tools:layout_conversion_absoluteWidth="0dp"
+                android:layout_row="0"
+                android:layout_column="49" />
+
+        </GridLayout>
+
     </android.support.v7.widget.Toolbar>
 
     <!--Títulos-->
