+++ app/src/main/res/layout/activity_method_viii.xml
@@ -1,54 +1,78 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
 
     tools:context="methods.MethodVIIIActivity">
 
     <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
+        android:id="@+id/toolbar2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/green_unifesp"
         android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
+        android:theme="?attr/actionBarTheme">
 
-        <RelativeLayout
+        <GridLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            android:columnCount="10"
+            android:rowCount="1">
+
+            <Button
+                android:id="@+id/button_goTests"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="end"
+
+                android:background="@drawable/custom_buttom"
+
+                android:textAllCaps="true"
+                android:textSize="18dp"
+                android:textStyle="bold"
+                android:text="Voltar"/>
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_gravity="center"
+                android:layout_marginLeft="210dp"
+                android:contentDescription="unifesp_escrito"
+                android:src="@drawable/unifesp_escrito" />
 
             <Button
                 android:id="@+id/button_goMenu"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:background="@drawable/custom_buttom"
-                android:text="Menu"
-                android:layout_alignParentStart="true"
-                android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
+                android:layout_gravity="end"
 
-            <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
-                android:src="@drawable/unifesp_escrito" />
-
-            <Button
-                android:id="@+id/button_goMethodIX"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
                 android:background="@drawable/custom_buttom"
-                android:text="Salvar e Continuar"
+
                 android:textAllCaps="true"
                 android:textSize="18dp"
-                android:textStyle="bold" />
-        </RelativeLayout>
+                android:textStyle="bold"
+                android:text="Menu"
+                android:layout_column="9"/>
+
+            <Button
+                android:id="@+id/button_goMethodIII"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+
+                android:background="@drawable/custom_buttom"
+
+                android:textAllCaps="true"
+                android:textSize="18dp"
+                android:textStyle="bold"
+                android:text="Próximo"
+                android:layout_column="9"/>
+
+        </GridLayout>
+
     </android.support.v7.widget.Toolbar>
 
     <TextView
