+++ app/src/main/res/layout/activity_method_xii.xml
@@ -6,49 +6,59 @@
     android:layout_height="match_parent"
     android:layout_marginBottom="8dp"
     android:orientation="vertical"
-    tools:context=".methods.MethodXIActivity">
+    tools:context=".methods.MethodXIIActivity">
 
     <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
+        android:id="@+id/toolbar2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/green_unifesp"
         android:minHeight="?attr/actionBarSize"
-        app:theme="?attr/actionBarTheme">
+        android:theme="?attr/actionBarTheme">
 
-        <RelativeLayout
+        <GridLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            android:columnCount="3"
+            android:rowCount="1">
 
             <Button
-                android:id="@+id/button_goMenu"
+                android:id="@+id/button_goMethodX"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:layout_gravity="end"
+
                 android:background="@drawable/custom_buttom"
-                android:text="Menu"
-                android:layout_alignParentStart="true"
+
                 android:textAllCaps="true"
-                android:textSize="18sp"
-                android:textStyle="bold" />
+                android:textSize="18dp"
+                android:textStyle="bold"
+                android:text="Voltar"/>
 
             <ImageView
-                android:layout_width="146dp"
-                android:layout_height="45dp"
-                android:layout_centerInParent="true"
-                android:contentDescription="Unifesp_Escrito"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_gravity="center"
+                android:layout_marginLeft="210dp"
+                android:contentDescription="unifesp_escrito"
                 android:src="@drawable/unifesp_escrito" />
 
             <Button
-                android:id="@+id/button_goEnd"
+                android:id="@+id/button_goMethodXII"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_alignParentEnd="true"
+                android:layout_gravity="end"
+
                 android:background="@drawable/custom_buttom"
-                android:text="Salvar e Continuar"
+
+                android:text="Terminar"
                 android:textAllCaps="true"
                 android:textSize="18dp"
                 android:textStyle="bold" />
-        </RelativeLayout>
+
+        </GridLayout>
+
     </android.support.v7.widget.Toolbar>
 
     <android.support.constraint.ConstraintLayout
